診断集約の総和をPRECによらずreal64に固定
calcstat の hsum/hsum_j/hgsum/hgsum_j と par_sum_rows(serial/mpi)の
引数を real(real64) 明示にする。単精度ビルドでも総和の桁落ちを防ぐ
(全国級 1e7 セルの総和は単精度だと有効桁を数桁失う)。状態量・出力量
は PREC のまま=「状態は PREC、診断集約は real64」の線引き。

倍精度ビルドでは real64=既定 real なので単一実装のまま従来とビット
同一(検証: PREC=double で逐次・np=1,2,4 が既存 reference と一致)。
par_sum_rows の MPI 通信型はバッファ型と常に一致する MPI_REAL8 固定に
変更(MPI_WP 規約の例外としてヘッダに明記)。§1 に方針の例外項を追加。

+++ src/m_parallel_serial.f90
@@ -22,7 +22,7 @@ module m_parallel
 !       do j = max(dcp%js, dcp%jw1+1), min(dcp%je, dcp%jw2-1)
 !     と書くこと(dcp%js+1 と書くと分割後にランク境界まで縮んでしまう)
 !=====================================================================
-   use, intrinsic :: iso_fortran_env, only: output_unit, error_unit
+   use, intrinsic :: iso_fortran_env, only: output_unit, error_unit, real64
    implicit none
    private
    public :: par_init, par_finalize
@@ -125,8 +125,10 @@ contains
    subroutine par_sum_rows(rowsum, total)
       ! 行部分和の総和。逐次では従来どおり一括総和する
       ! (MPI 版も rank0 が同じ「全窓行の一括総和」を行う)。
-      real, intent(in) :: rowsum(dcp%js:)
-      real, intent(out) :: total
+      ! 引数は PREC によらず real64 固定(診断集約の桁落ち防止。
+      ! 倍精度ビルドでは real64=既定 real なので単一実装で済む。§1)
+      real(real64), intent(in) :: rowsum(dcp%js:)
+      real(real64), intent(out) :: total
       total = sum(rowsum)
    end subroutine par_sum_rows
